+++ packages/events/src/firebase.ts
@@ -16,3 +16,5 @@ export * from '@cloudcommerce/app-infinitepay/webhook';
 export * from '@cloudcommerce/app-galaxpay/events';
 
 export * from '@cloudcommerce/app-loyalty-points/events';
+
+export * from '@cloudcommerce/app-frenet/events';