+++ packages/storefront/src/lib/pages/_vue.ts
@@ -1,6 +1,5 @@
 import type { App } from 'vue';
 import { i18n, formatMoney } from '@ecomplus/utils';
-import VTooltip from 'floating-vue/src/directives/v-tooltip';
 // @ts-ignore
 import Fade from '@@components/globals/Fade.vue';
 
@@ -21,7 +20,6 @@ export default (app: App) => {
       app.config.globalProperties.$percentage = formatPercentage;
     },
   });
-  app.directive('tooltip', VTooltip);
   app.component('Fade', Fade);
 };