+++ packages/storefront/tailwind.config.cjs
@@ -37,6 +37,8 @@ let defaultThemeOptions = {
     'dinersclub',
   ],
   generalIcons: 'heroicons',
+  shoppingCartIcon: 'shopping-bag',
+  cashbackIcon: 'arrow-uturn-left',
 };
 if (globalThis.storefront_theme_options) {
   defaultThemeOptions = deepmerge(defaultThemeOptions, globalThis.storefront_theme_options);
@@ -102,17 +104,20 @@ Object.keys(brandColors).forEach((colorName) => {
   });
 });
 
-const genTailwindConfig = ({
-  brandIcons,
-  brandIconsShortcuts,
-  brandLogos,
-  brandLogosShortcuts,
-  generalIcons,
-  baseColor,
-  successColor,
-  warningColor,
-  dangerColor,
-} = defaultThemeOptions) => {
+const genTailwindConfig = (themeOptions = {}) => {
+  const {
+    brandIcons,
+    brandIconsShortcuts,
+    brandLogos,
+    brandLogosShortcuts,
+    generalIcons,
+    baseColor,
+    successColor,
+    warningColor,
+    dangerColor,
+    shoppingCartIcon,
+    cashbackIcon,
+  } = deepmerge(defaultThemeOptions, themeOptions);
   const config = {
     theme: {
       extend: {
@@ -157,9 +162,8 @@ const genTailwindConfig = ({
               if (!shortcuts) {
                 const { icons } = require(`@iconify-json/${iconset}`);
                 shortcuts = Object.keys(icons.icons);
-                if (!shortcuts.includes('shopping-cart')) {
-                  shortcuts.push('shopping-cart');
-                }
+                shortcuts.push(['shopping-cart', shoppingCartIcon]);
+                shortcuts.push(['cashback', cashbackIcon]);
               }
               shortcuts.forEach((shortcut) => {
                 if (typeof shortcut === 'string') {