+++ scripts/pre-install.sh
@@ -1,0 +1,10 @@
+#!/bin/bash
+
+rm -rf store/node_modules
+rm -rf store/package-lock.json
+rm -rf store/functions/*/node_modules
+rm -rf store/functions/*/package-lock.json
+rm -rf ecomplus-stores/*/node_modules
+rm -rf ecomplus-stores/*/package-lock.json
+rm -rf ecomplus-stores/*/functions/*/node_modules
+rm -rf ecomplus-stores/*/functions/*/package-lock.json